security/pyproject.toml
[project]
name = "demiurge"
version = "0.1.0"
description = "Demiurge — orchestrator + sole broker for secrets and sensitive operations across the Pantheon"
requires-python = ">=3.10"
dependencies = [
    "shared",
    "msgpack>=1.0",
    "pynacl>=1.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "keyring>=24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
janus = [
    # Playwright-driven operator-assisted browser automation.
    # Operators who never run `demiurge janus` don't pay the install cost.
    # Install: `uv sync --extra janus` then `uv run playwright install chromium`.
    "playwright>=1.40",
]

[project.scripts]
demiurge = "demiurge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/demiurge"]
